Packaging and CI updates
- Support Python 3.9+
- Use `tox-uv`
- Build on arm-based mac runners
- Add dependabot updates for github actions
- Use hatch-based packaging
