+++ .github/workflows/ci.yml
@@ -17,8 +17,8 @@ jobs:
     tests:
         strategy:
             matrix:
-                python_version: ["3.7", "3.9", "3.10", "3.11"]
-                os: [ubuntu-latest, windows-latest, macos-latest]
+                python_version: ["3.9", "3.10", "3.11", "3.12"]
+                os: [ubuntu-latest, windows-latest, macos-14]
         runs-on: ${{ matrix.os }}
         steps:
             - uses: actions/checkout@v3
@@ -28,7 +28,7 @@ jobs:
                 python-version: '${{ matrix.python_version }}'
                 cache: 'pip'
             - name: install tox
-              run: python3 -m pip install tox
+              run: python3 -m pip install tox tox-uv
             - name: run tox
               run: tox -e 'py${{ matrix.python_version }}'
             - name: submit code coverage
@@ -104,10 +104,10 @@ jobs:
         LT_RELEASE_UPLOAD_ZENODO_PARENT: ${{ secrets.LT_RELEASE_UPLOAD_ZENODO_PARENT }}
       steps:
       - uses: actions/checkout@v3
-      - name: Choose Python version 3.10
+      - name: Choose Python version 3.11
         uses: actions/setup-python@v4
         with:
-          python-version: "3.10"
+          python-version: "3.11"
           cache: 'pip'
       - name: install release script deps
         run: pip install -r scripts/requirements.txt
