--- MANIFEST.in
@@ -1,3 +1,0 @@
-graft src
-include LICENSE
-include README.rst
