--- appveyor.yml
@@ -1,26 +1,0 @@
-build: false
-environment:
-  matrix:
-    - PYTHON: "C:\\Python36-x64"
-      PYTHON_VERSION: "3.6.5"
-      PYTHON_ARCH: "64"
-      TOXENV: "py36"
-
-    - PYTHON: "C:\\Python37-x64"
-      PYTHON_VERSION: "3.7"
-      PYTHON_ARCH: "64"
-      TOXENV: "py37"
-
-init:
-  - "ECHO %PYTHON% %PYTHON_VERSION% %PYTHON_ARCH%"
-
-install:
-  - "%PYTHON%/Scripts/pip.exe install --force-reinstall tox codecov"
-
-test_script:
-  - "%PYTHON%/Scripts/tox"
-
-on_success:
-  - "%PYTHON%/Scripts/codecov -f coverage.xml"
-
-skip_tags: true
