+++ setup.cfg
@@ -1,9 +1,3 @@
-[bdist_wheel]
-universal = 1
-
-[metadata]
-license_file = LICENSE
-
 [flake8]
 max-line-length = 100
 ignore = E24,E121,E123,E126,E128,E133,E226,E241,E242,E704,W503
