+++ tox.ini
@@ -1,5 +1,5 @@
 [tox]
-envlist = py{37,38,39,310,311}
+envlist = py{39,310,311,312}
 
 [testenv]
 commands=
